Strongly_Connected_Components
Strongly connected labels and Weakly connected labels are part of a  graph. Either it is directed or undirected graphs.

Hello Everyone.

We all know that the Graph is used in Mapping analysis, Short distance measurements, Data Management, and much more.

So, I have made this Program in Python language in which you can find the connected components more precisely the most strongly connected components.

In the code, I have used the DFS Algorithm to find the connected components (nodes). The program will make a text file for the connected and least connected nodes present in a graph.

#Note

You Need a Good GPU to run this code for a massive number of nodes in a graph.

Here is the link for this respective code